Rename builder/animation_combinator module to time
